fix(capdir): read a device number the same way on every platform's width
`st_dev` is an i32 on macOS and a u64 on Linux, so `u64::try_from` was a conversion on one
and the identity on the other, which CI's Linux clippy refused.
Going through `i128::from`, lossless from every width the field has, makes the checked step
a conversion on both; `mise run lint:linux` reads it clean.

+++ crates/rust-mutants/src/capdir/unix.rs
@@ -79,7 +79,7 @@ fn status_of(stat: &rustix::fs::Stat) -> io::Result<Status> {
         | FileType::BlockDevice
         | FileType::Unknown => Kind::Other,
     };
-    let volume = u64::try_from(stat.st_dev)
+    let volume = u64::try_from(i128::from(stat.st_dev))
         .map_err(|_negative| io::Error::other("a device number is negative"))?;
     let object = u128::from(stat.st_ino);
     let len = u64::try_from(stat.st_size)
